+++ samples/TorProxy.Sandbox/Program.cs
@@ -24,10 +24,6 @@ namespace Nefarius.Utilities.TorProxy.Sandbox
             {
                 ZippedToolsDirectory = Path.Combine(Path.GetTempPath(), "TorZipped"),
                 ExtractedToolsDirectory = Path.Combine(Path.GetTempPath(), "TorExtracted"),
-                PrivoxySettings =
-                {
-                    Port = 18118,
-                },
                 TorSettings =
                 {
                     SocksPort = 19050,
@@ -51,9 +47,12 @@ namespace Nefarius.Utilities.TorProxy.Sandbox
             {
                 var fetcher = new TorProxyToolFetcher(settings, httpClient);
                 var updates = await fetcher.CheckForUpdatesAsync();
-                Console.WriteLine($"Current Privoxy: {updates.Privoxy.LocalVersion?.ToString() ?? "(none)"}");
-                Console.WriteLine($" Latest Privoxy: {updates.Privoxy.LatestDownload.Version}");
-                Console.WriteLine();
+                if (updates.Privoxy != null)
+                {
+                    Console.WriteLine($"Current Privoxy: {updates.Privoxy.LocalVersion?.ToString() ?? "(none)"}");
+                    Console.WriteLine($" Latest Privoxy: {updates.Privoxy.LatestDownload.Version}");
+                    Console.WriteLine();
+                }
                 Console.WriteLine($"Current Tor: {updates.Tor.LocalVersion?.ToString() ?? "(none)"}");
                 Console.WriteLine($" Latest Tor: {updates.Tor.LatestDownload.Version}");
                 Console.WriteLine();
@@ -68,7 +67,7 @@ namespace Nefarius.Utilities.TorProxy.Sandbox
             {
                 var handler = new HttpClientHandler
                 {
-                    Proxy = new WebProxy(new Uri("http://localhost:" + settings.PrivoxySettings.Port))
+                    Proxy = new WebProxy(new Uri("socks5://localhost:" + settings.TorSettings.SocksPort))
                 };
 
                 using (handler)